GW-020 optimizied imports + minor renaming of files [TA, WF]
