+++ app/src/androidTest/java/at/tugraz/ist/guessingwords/MultiplayerActivityTest.kt
@@ -5,7 +5,6 @@ import androidx.test.espresso.Espresso
 import androidx.test.espresso.action.ViewActions
 import androidx.test.espresso.assertion.ViewAssertions
 import androidx.test.espresso.matcher.ViewMatchers
-import androidx.test.platform.app.InstrumentationRegistry
 import org.junit.Test
 
 class MultiplayerActivityTest {
