+++ app/src/androidTest/java/at/tugraz/ist/guessingwords/logic/GameTest.kt
@@ -4,7 +4,6 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import at.tugraz.ist.guessingwords.data.entity.Word
 import org.junit.Test
 import org.junit.runner.RunWith
-import java.lang.IllegalArgumentException
 
 @RunWith(AndroidJUnit4::class)
 class GameTest {
