+++ app/src/main/java/at/tugraz/ist/guessingwords/CustomWordsActivity.kt
@@ -1,8 +1,8 @@
 package at.tugraz.ist.guessingwords
 
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.MenuItem
+import androidx.appcompat.app.AppCompatActivity
 
 
 class CustomWordsActivity : AppCompatActivity() {
