+++ app/src/main/java/at/tugraz/ist/guessingwords/HostActivity.kt
@@ -1,11 +1,8 @@
 package at.tugraz.ist.guessingwords
 
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.MenuItem
-import android.widget.Button
-import android.widget.Toast
+import androidx.appcompat.app.AppCompatActivity
 
 class HostActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
