+++ app/src/main/java/at/tugraz/ist/guessingwords/JoinActivity.kt
@@ -1,8 +1,8 @@
 package at.tugraz.ist.guessingwords
 
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.MenuItem
+import androidx.appcompat.app.AppCompatActivity
 
 class JoinActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
