+++ app/src/main/java/at/tugraz/ist/guessingwords/MultiplayerActivity.kt
@@ -1,11 +1,10 @@
 package at.tugraz.ist.guessingwords
 
 import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.MenuItem
 import android.widget.Button
-import android.widget.Toast
+import androidx.appcompat.app.AppCompatActivity
 
 class MultiplayerActivity : AppCompatActivity() {
