+++ app/src/test/java/at/tugraz/ist/guessingwords/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
 package at.tugraz.ist.guessingwords
 
+import org.junit.Assert.assertEquals
 import org.junit.Test
-
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
